+++ ircd/class_ext.h
@@ -39,11 +39,11 @@ EXTERN int get_client_class __P((aClient *acptr));
 EXTERN int get_client_ping __P((aClient *acptr));
 EXTERN int get_con_freq __P((aClass *clptr));
 EXTERN void add_class __P((int class, int ping, int confreq, int maxli,
-			   long sendq, int hlocal, int uhlocal,
+			   int sendq, int hlocal, int uhlocal,
 			   int hglobal, int uhglobal));
 EXTERN aClass *find_class __P((int cclass));
 EXTERN void check_class();
 EXTERN void initclass();
 EXTERN void report_classes __P((aClient *sptr, char *to));
-EXTERN long get_sendq __P((aClient *cptr));
+EXTERN int get_sendq __P((aClient *cptr));
 #undef EXTERN
